+++ tests/unit_tests/serial/run_panda_plans/test_panda_serial_collection.py
@@ -9,8 +9,6 @@ from i19_bluesky.serial.device_setup_plans.diffractometer_plans import (
     setup_sample_stage,
 )
 from i19_bluesky.serial.run_panda_plans.panda_serial_collection import (
-    end_run,
-    run_on_collection_abort,
     trigger_panda_collection,
 )
 
@@ -32,8 +30,6 @@ async def test_setup_sample_stage(
 @patch("i19_bluesky.serial.run_panda_plans.panda_serial_collection.bps.trigger")
 @patch("i19_bluesky.serial.run_panda_plans.panda_serial_collection.bps.abs_set")
 @patch("i19_bluesky.serial.run_panda_plans.panda_serial_collection.move_stage_x_and_z")
-@patch("i19_bluesky.serial.run_panda_plans.panda_serial_collection.reset_panda")
-@patch("i19_bluesky.serial.run_panda_plans.panda_serial_collection.disarm_panda")
 @patch("i19_bluesky.serial.run_panda_plans.panda_serial_collection.arm_panda")
 @patch(
     "i19_bluesky.serial.run_panda_plans.panda_serial_collection.setup_panda_for_rotation"
@@ -41,8 +37,6 @@ async def test_setup_sample_stage(
 def test_trigger_panda_call_order(
     mock_setup_panda_for_rotation: MagicMock,
     mock_arm_panda: MagicMock,
-    mock_disarm_panda: MagicMock,
-    mock_reset_panda: MagicMock,
     mock_move_stage_x_and_z: MagicMock,
     mock_set_value_for_params: MagicMock,
     mock_arm_or_disarm: MagicMock,
@@ -59,8 +53,6 @@ def test_trigger_panda_call_order(
         mock_setup_panda_for_rotation, "mock_setup_panda_for_rotation"
     )
     parent_mock.attach_mock(mock_arm_panda, "mock_arm_panda")
-    parent_mock.attach_mock(mock_disarm_panda, "mock_disarm_panda")
-    parent_mock.attach_mock(mock_reset_panda, "mock_reset_panda")
     parent_mock.attach_mock(mock_arm_or_disarm, "mock_arm_or_disarm")
     RE(trigger_panda_collection(parameters, devices))
     expected_calls = [
@@ -75,46 +67,3 @@ def test_trigger_panda_call_order(
         call.mock_set_value_for_params(devices.diffractometer.phi, 5.0, wait=True),
     ]
     parent_mock.assert_has_calls(expected_calls, any_order=True)
-
-
-@patch("i19_bluesky.serial.run_panda_plans.panda_serial_collection.reset_panda")
-@patch(
-    "i19_bluesky.serial.run_panda_plans.panda_serial_collection.move_sample_stage_back"
-)
-@patch("i19_bluesky.serial.run_panda_plans.panda_serial_collection.bps.trigger")
-@patch("i19_bluesky.serial.run_panda_plans.panda_serial_collection.disarm_panda")
-async def test_end_run(
-    mock_disarm_panda: MagicMock,
-    mock_disarm_eiger: MagicMock,
-    mock_move_sample_stage_back: MagicMock,
-    mock_reset_panda: MagicMock,
-    parameters: SerialExperimentEh2,
-    devices: SerialCollectionEh2PandaComposite,
-    RE: RunEngine,
-):
-    RE(
-        end_run(
-            parameters.rot_axis_start,
-            devices.panda,
-            devices.eiger,
-            devices.serial_stages,
-        )
-    )
-    mock_disarm_eiger.assert_called_once_with(devices.eiger.detector.disarm)
-    mock_move_sample_stage_back.assert_called_once_with(devices.serial_stages, 0)
-    mock_disarm_panda.assert_called_once_with(devices.panda)
-    mock_reset_panda.assert_called_once_with(devices.panda)
-
-
-@patch("i19_bluesky.serial.run_panda_plans.panda_serial_collection.bps.trigger")
-@patch("i19_bluesky.serial.run_panda_plans.panda_serial_collection.disarm_panda")
-async def test_run_on_collection_abort(
-    mock_disarm_panda: MagicMock,
-    mock_disarm_eiger: MagicMock,
-    RE: RunEngine,
-    devices: SerialCollectionEh2PandaComposite,
-):
-    RE(run_on_collection_abort(devices.panda, devices.eiger, devices.diffractometer))
-    get_mock_put(devices.diffractometer.phi.motor_stop).assert_called_once_with(1)
-    mock_disarm_eiger.assert_called_once_with(devices.eiger.detector.disarm)
-    mock_disarm_panda.assert_called_once_with(devices.panda)
