+++ tests/unit_tests/serial/run_panda_plans/test_run_serial_with_panda.py
@@ -1,16 +1,19 @@
 from unittest.mock import MagicMock, patch
 
 from bluesky.run_engine import RunEngine
+from ophyd_async.core import get_mock_put
 
 from i19_bluesky.parameters.devices_composites import SerialCollectionEh2PandaComposite
 from i19_bluesky.parameters.serial_parameters import SerialExperimentEh2
 from i19_bluesky.serial.run_panda_plans.run_serial_with_panda import (
     main_collection_plan,
+    run_on_collection_abort,
+    run_on_collection_end,
     run_serial_with_panda,
 )
 
 
-@patch("i19_bluesky.serial.run_panda_plans.run_serial_with_panda.end_run")
+@patch("i19_bluesky.serial.run_panda_plans.run_serial_with_panda.run_on_collection_end")
 @patch("i19_bluesky.serial.run_panda_plans.run_serial_with_panda.main_collection_plan")
 async def test_run_serial_with_panda(
     mock_main_plan: MagicMock,
@@ -46,3 +49,46 @@ async def test_main_collection_plan(
     RE(main_collection_plan(parameters, devices))
     mock_setup_collection.assert_called_once_with(parameters, devices)
     mock_trigger_panda.assert_called_once_with(parameters, devices)
+
+
+@patch("i19_bluesky.serial.run_panda_plans.run_serial_with_panda.bps.trigger")
+@patch("i19_bluesky.serial.run_panda_plans.run_serial_with_panda.disarm_panda")
+async def test_run_on_collection_abort(
+    mock_disarm_panda: MagicMock,
+    mock_disarm_eiger: MagicMock,
+    RE: RunEngine,
+    devices: SerialCollectionEh2PandaComposite,
+):
+    RE(run_on_collection_abort(devices.panda, devices.eiger, devices.diffractometer))
+    get_mock_put(devices.diffractometer.phi.motor_stop).assert_called_once_with(1)
+    mock_disarm_eiger.assert_called_once_with(devices.eiger.detector.disarm)
+    mock_disarm_panda.assert_called_once_with(devices.panda)
+
+
+@patch("i19_bluesky.serial.run_panda_plans.run_serial_with_panda.reset_panda")
+@patch(
+    "i19_bluesky.serial.run_panda_plans.run_serial_with_panda.move_sample_stage_back"
+)
+@patch("i19_bluesky.serial.run_panda_plans.run_serial_with_panda.bps.trigger")
+@patch("i19_bluesky.serial.run_panda_plans.run_serial_with_panda.disarm_panda")
+async def test_end_run(
+    mock_disarm_panda: MagicMock,
+    mock_disarm_eiger: MagicMock,
+    mock_move_sample_stage_back: MagicMock,
+    mock_reset_panda: MagicMock,
+    parameters: SerialExperimentEh2,
+    devices: SerialCollectionEh2PandaComposite,
+    RE: RunEngine,
+):
+    RE(
+        run_on_collection_end(
+            parameters.rot_axis_start,
+            devices.panda,
+            devices.eiger,
+            devices.serial_stages,
+        )
+    )
+    mock_disarm_eiger.assert_called_once_with(devices.eiger.detector.disarm)
+    mock_move_sample_stage_back.assert_called_once_with(devices.serial_stages, 0)
+    mock_disarm_panda.assert_called_once_with(devices.panda)
+    mock_reset_panda.assert_called_once_with(devices.panda)
